+++ src/sao.c
@@ -423,8 +423,8 @@ static void sao_search_edge_sao(const encoder_state_t * const state,
     }
 
     {
-      float mode_bits = sao_mode_bits_edge(state, edge_class, edge_offset, sao_top, sao_left, buf_cnt);
-      sum_ddistortion += (int)((double)mode_bits*state->lambda +0.5);
+      double mode_bits = sao_mode_bits_edge(state, edge_class, edge_offset, sao_top, sao_left, buf_cnt);
+      sum_ddistortion += (int)(mode_bits*state->lambda +0.5);
     }
     // SAO is not applied for category 0.
     edge_offset[SAO_EO_CAT0] = 0;
@@ -456,7 +456,7 @@ static void sao_search_band_sao(const encoder_state_t * const state, const uvg_p
     int sao_bands[2][32];
     int temp_offsets[10];
     int ddistortion = 0;
-    float temp_rate = 0.0;
+    double temp_rate = 0.0;
     
     for (i = 0; i < buf_cnt; ++i) {
       FILL(sao_bands, 0);
@@ -468,7 +468,7 @@ static void sao_search_band_sao(const encoder_state_t * const state, const uvg_p
     }
 
     temp_rate = sao_mode_bits_band(state, sao_out->band_position, temp_offsets, sao_top, sao_left, buf_cnt);
-    ddistortion += (int)((double)temp_rate*state->lambda + 0.5);
+    ddistortion += (int)(temp_rate*state->lambda + 0.5);
 
     // Select band sao over edge sao when distortion is lower
     if (ddistortion < sao_out->ddistortion) {
@@ -509,7 +509,7 @@ static void sao_search_best_mode(const encoder_state_t * const state, const uvg_
 
   if (state->encoder_control->cfg.sao_type & 1){
     sao_search_edge_sao(state, data, recdata, block_width, block_height, buf_cnt, &edge_sao, sao_top, sao_left);
-    float mode_bits = sao_mode_bits_edge(state, edge_sao.eo_class, edge_sao.offsets, sao_top, sao_left, buf_cnt);
+    double mode_bits = sao_mode_bits_edge(state, edge_sao.eo_class, edge_sao.offsets, sao_top, sao_left, buf_cnt);
     int ddistortion = (int)(mode_bits * state->lambda + 0.5);
     unsigned buf_i;
     
@@ -527,7 +527,7 @@ static void sao_search_best_mode(const encoder_state_t * const state, const uvg_
 
   if (state->encoder_control->cfg.sao_type & 2){
     sao_search_band_sao(state, data, recdata, block_width, block_height, buf_cnt, &band_sao, sao_top, sao_left);
-    float mode_bits = sao_mode_bits_band(state, band_sao.band_position, band_sao.offsets, sao_top, sao_left, buf_cnt);
+    double mode_bits = sao_mode_bits_band(state, band_sao.band_position, band_sao.offsets, sao_top, sao_left, buf_cnt);
     int ddistortion = (int)(mode_bits * state->lambda + 0.5);
     unsigned buf_i;
     
@@ -554,7 +554,7 @@ static void sao_search_best_mode(const encoder_state_t * const state, const uvg_
   // Choose between SAO and doing nothing, taking into account the
   // rate-distortion cost of coding do nothing.
   {
-    float mode_bits_none = sao_mode_bits_none(state, sao_top, sao_left);
+    double mode_bits_none = sao_mode_bits_none(state, sao_top, sao_left);
     int cost_of_nothing = (int)(mode_bits_none * state->lambda + 0.5);
     if (sao_out->ddistortion >= cost_of_nothing) {
       sao_out->type = SAO_TYPE_NONE;
@@ -571,7 +571,7 @@ static void sao_search_best_mode(const encoder_state_t * const state, const uvg_
 
       if (merge_cand) {
         unsigned buf_i;
-        float mode_bits = sao_mode_bits_merge(state, i + 1);
+        double mode_bits = sao_mode_bits_merge(state, i + 1);
         int ddistortion = (int)(mode_bits * state->lambda + 0.5);
 
         switch (merge_cand->type) {
